userclk: separate device independent from Agilex 7 configuration
The user clock frequency configuration protocol varies across devices.
Move the Agilex 7 and Stratix 10 configuration sequences out of
fpga_user_clk.c and into a new protocol-specific file.
